+++ src/main/java/io/github/galop_proxy/galop/configuration/ConfigurationPropertyKeys.java
@@ -3,6 +3,8 @@ package io.github.galop_proxy.galop.configuration;
 final class ConfigurationPropertyKeys {
 
     static final String PROXY_PORT = "proxy.port";
+    static final String PROXY_BACKLOG_SIZE = "proxy.backlog_size";
+    static final String PROXY_BIND_ADDRESS = "proxy.bind_address";
 
     static final String TARGET_ADDRESS = "target.address";
     static final String TARGET_PORT = "target.port";