+++ src/main/java/io/github/galop_proxy/galop/configuration/HttpHeaderTypeConfiguration.java
@@ -1,0 +1,9 @@
+package io.github.galop_proxy.galop.configuration;
+
+interface HttpHeaderTypeConfiguration {
+
+    long getReceiveTimeout();
+
+    int getMaxSize();
+
+}