+++ src/main/java/io/github/galop_proxy/galop/http/HttpConstants.java
@@ -8,13 +8,13 @@ class HttpConstants {
 
     static final Charset HEADER_CHARSET = Charset.forName("ASCII");
 
-    static final String HEADER_HOST_PREFIX = "Host:".toLowerCase();
-    static final String HEADER_DATE_PREFIX = "Date:".toLowerCase();
-    static final String HEADER_SERVER_PREFIX = "Server:".toLowerCase();
-    static final String HEADER_CONTENT_LENGTH_PREFIX = "Content-Length:".toLowerCase();
-    static final String HEADER_CONTENT_TYPE_PREFIX = "Content-Type:".toLowerCase();
+    static final String HEADER_HOST_PREFIX = "host:";
+    static final String HEADER_DATE_PREFIX = "date:";
+    static final String HEADER_SERVER_PREFIX = "server:";
+    static final String HEADER_CONTENT_LENGTH_PREFIX = "content-length:";
+    static final String HEADER_CONTENT_TYPE_PREFIX = "content-type:";
     static final String HEADER_CONTENT_TYPE_CHARSET_PREFIX = "charset=";
-    static final String HEADER_TRANSFER_ENCODING_PREFIX = "Transfer-Encoding:".toLowerCase();
+    static final String HEADER_TRANSFER_ENCODING_PREFIX = "transfer-encoding:";
 
     static final String TRANSFER_ENCODING_IDENTITY = "identity";
     static final String TRANSFER_ENCODING_CHUNKED = "chunked";