+++ src/test/java/io/github/galop_proxy/galop/configuration/ProxyConfigurationImplTest.java
@@ -4,7 +4,11 @@ import io.github.galop_proxy.galop.commons.PortNumber;
 import org.junit.Before;
 import org.junit.Test;
 
+import java.net.InetAddress;
+
 import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertNull;
+import static org.mockito.Mockito.mock;
 
 /**
  * Tests the class {@link ProxyConfigurationImpl}.
@@ -12,12 +16,16 @@ import static org.junit.Assert.assertEquals;
 public class ProxyConfigurationImplTest {
 
     private PortNumber port;
+    private int backlogSize;
+    private InetAddress bindAddress;
     private ProxyConfiguration configuration;
 
     @Before
     public void setUp() {
         port = new PortNumber(8080);
-        configuration = new ProxyConfigurationImpl(port);
+        backlogSize = 50;
+        bindAddress = mock(InetAddress.class);
+        configuration = new ProxyConfigurationImpl(port, backlogSize, bindAddress);
     }
 
     @Test
@@ -25,9 +33,25 @@ public class ProxyConfigurationImplTest {
         assertEquals(port, configuration.getPort());
     }
 
+    @Test
+    public void getBacklogSize_returnsConfiguredBacklogSize() {
+        assertEquals(backlogSize, configuration.getBacklogSize());
+    }
+
+    @Test
+    public void getBindAddress_returnsConfiguredInetAddress() {
+        assertEquals(bindAddress, configuration.getBindAddress());
+    }
+
+    @Test
+    public void getBindAddress_whenNoBindAddressIsConfigured_returnsNull() {
+        configuration = new ProxyConfigurationImpl(port, backlogSize, null);
+        assertNull(configuration.getBindAddress());
+    }
+
     @Test(expected = NullPointerException.class)
     public void constructor_withoutPort_throwsNullPointerException() {
-        new ProxyConfigurationImpl(null);
+        new ProxyConfigurationImpl(null, backlogSize, null);
     }
 
 }